Add files via upload
This project was done using html, css and js, and it is a responsive design that works perfectly on all devices and screens. The goal of the project is to apply it to the Javascript language in order to improve the skill in it due to its importance to the web front-end developer.